Move compute_auth_events to EventAuthHandler.
